[REFACTOR] github login에 필요한 data class feature -> domain 모듈로 이동 (#17)
