+++ domain/build.gradle.kts
@@ -7,3 +7,7 @@ java {
     sourceCompatibility = JavaVersion.VERSION_17
     targetCompatibility = JavaVersion.VERSION_17
 }
+
+dependencies {
+    api(libs.bundles.gson)
+}
