+++ feature/auth/src/main/java/com/lgtm/android/auth/ui/SignInViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.google.gson.Gson
-import com.lgtm.android.auth.model.GithubLoginResponse
+import com.lgtm.domain.entity.response.GithubLoginResponse
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
